feat(getting-started-2.tsx): add screen for user to input their name and save it to local storage
feat(getting-started): add GettingStarted3 screen and refactor GettingStarted screen

Add a new screen GettingStarted3 that displays an image and a call to action button. Refactor GettingStarted screen to use a custom button component and to navigate to GettingStarted2 screen when the button is pressed. Also, change the background color of GettingStarted screen to match the new screen.