feat(getting-started): add support for adding bank account name
Add support for adding bank account name in the getting started flow. The
`handleName` function has been moved to a separate file `handleName.tsx` to
make it reusable. The `GettingStarted2` screen now uses this function to
validate the name input. The `GettingStarted3` screen now allows the user to
input the account name and saves it to the device using `expo-secure-store`.
The `package.json` file has been updated to include `expo-sqlite` and
`expo-file-system` as dependencies. The `getting-started-account.png` image
has been added to the assets folder. Some minor style changes have also been
made to the screens.