style(ui/button.tsx, getting-started-2.tsx, getting-started-3.tsx, getting-started.tsx): change background color of some components
Change the background color of the primary button in ui/button.tsx from #06D188 to #41BB82. In getting-started-2.tsx, change the background color of the main container from #f0fdf4 to #F0713603 and the background color of the primary button to #F07136. In getting-started-3.tsx, change the background color of the main container from #f0fdf4 to #D181B603 and the background color of the primary button to #D181B6. In getting-started.tsx, change the background color of the main container from #f0fdf4 to #41BB821A.